chore(rename): rename tickables to triggers
lol test_tickables.py
