+++ src/froggytv/triggers.py
